permission fix. Scripts updated

+++ SNDG/Comparative/MultiRefCoreAln.py
@@ -43,20 +43,21 @@ class MultiRefCoreAln:
                     upair2=cepa_reads_core_r1, upair=cepa_reads_core_r2, ubam=aln_pe_bam)
             if os.path.getsize(cepa_reads_core_r1) < 1000:
                 print("NOOOOOOOOO!!!!!!!!!!!!!!!!!!!: " + cepa + " -> " + ref)
-            
-            execute(
-                'bwa mem -t {cpus} -R "@RG\\tID:illumina\\tSM:{ncepa}\\tLB:{ncepa}"  {ref_fasta} {s1}   >  ' + aln_s_sam,
-                ref_fasta=ref_fasta, ncepa=cepa, s1=in_fasta_s, cpus=self.cpus)
-            execute("samtools view -@ %s -F 4 -Sbh %s > %s" % (self.cpus, aln_s_sam, aln_s_bam))
-            os.remove(aln_s_sam)
-            execute("bedtools bamtofastq -i {ubam} -fq {upair} ", upair=cepa_reads_core_s, ubam=aln_s_bam)
+            if in_fasta_s:
+                execute(
+                    'bwa mem -t {cpus} -R "@RG\\tID:illumina\\tSM:{ncepa}\\tLB:{ncepa}"  {ref_fasta} {s1}   >  ' + aln_s_sam,
+                    ref_fasta=ref_fasta, ncepa=cepa, s1=in_fasta_s, cpus=self.cpus)
+                execute("samtools view -@ %s -F 4 -Sbh %s > %s" % (self.cpus, aln_s_sam, aln_s_bam))
+                os.remove(aln_s_sam)
+                execute("bedtools bamtofastq -i {ubam} -fq {upair} ", upair=cepa_reads_core_s, ubam=aln_s_bam)
+                in_fasta_s = cepa_reads_core_s
+                shutil.copy(cepa_reads_core_s, cepa_reads_core_s_bk)
+                
             in_fasta_r1 = cepa_reads_core_r1
             in_fasta_r2 = cepa_reads_core_r2
-            in_fasta_s = cepa_reads_core_s
-            shutil.copy(cepa_reads_core_s, cepa_reads_core_s_bk)
-
             shutil.copy(cepa_reads_core_r1, cepa_reads_core_r1_bk)
             shutil.copy(cepa_reads_core_r2, cepa_reads_core_r2_bk)
+            
         for x in os.listdir(output):
             if x.endswith(".bk"):
                 os.remove(x)
@@ -64,13 +65,15 @@ class MultiRefCoreAln:
                 os.remove(x)
         execute(f'gzip {cepa_reads_core_r1}')
         execute(f'gzip {cepa_reads_core_r2}')
-        execute(f'gzip {cepa_reads_core_s}')
+        if in_fasta_s:
+            execute(f'gzip {cepa_reads_core_s}')
 
 
 if __name__ == '__main__':
     import argparse
     import os
     from glob import glob
+    from SNDG import mkdir
 
     parser = argparse.ArgumentParser(description='Mapping to variant calls pipeline.')
     required = parser.add_argument_group('required arguments')
@@ -78,7 +81,7 @@ if __name__ == '__main__':
     required.add_argument('-o', '--output', required=True)
 
     required.add_argument('-s', '--sample_name', required=True)
-    required.add_argument('-rs', '--in_fasta_s', required=True)
+    required.add_argument('-rs', '--in_fasta_s', default=None)
     required.add_argument('-r2', '--in_fasta_r2', required=True)
     required.add_argument('-r1', '--in_fasta_r1', required=True)
     required.add_argument('-t', '--tmp_dir', default="/tmp")
@@ -88,7 +91,7 @@ if __name__ == '__main__':
 
     if not os.path.exists(args.ref_dirs):
         raise FileNotFoundError(f"{args.ref_dirs} does not exists")
-    if not os.path.exists(args.in_fasta_s):
+    if args.in_fasta_s and not os.path.exists(args.in_fasta_s):
         raise FileNotFoundError(f"{args.in_fasta_s} does not exists")
     if not os.path.exists(args.in_fasta_r2):
         raise FileNotFoundError(f"{args.in_fasta_r2} does not exists")
@@ -100,6 +103,13 @@ if __name__ == '__main__':
     if not refs:
         raise FileNotFoundError(f"no references detected in {args.ref_dirs}")
 
+    mkdir(args.output)
+    if not os.path.exists(args.output):
+        raise FileNotFoundError(f"could not create {args.output}")
+    
+    if not args.output.endswith("/"):
+        args.output = args.output + "/"
+    
     mrca = MultiRefCoreAln(base_refs=refs, cpus=args.cpus)
     mrca.core_reads(args.output, args.sample_name,
                     args.in_fasta_r1, args.in_fasta_r2, args.in_fasta_s, args.tmp_dir)
